# lib/db/src/schema/generated_images.ts
import { pgTable, serial, text, integer, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const generatedImages = pgTable("generated_images", {
  id: serial("id").primaryKey(),
  prompt: text("prompt").notNull(),
  url: text("url").notNull(),
  width: integer("width").notNull().default(1024),
  height: integer("height").notNull().default(1024),
  model: text("model").notNull().default("flux"),
  createdAt: timestamp("created_at").notNull().defaultNow(),
});

export const insertGeneratedImageSchema = createInsertSchema(generatedImages).omit({ id: true, createdAt: true });
export type InsertGeneratedImage = z.infer<typeof insertGeneratedImageSchema>;
export type GeneratedImage = typeof generatedImages.$inferSelect;